Move sharing from pkg to model
