+++ model/sharing/rule.go
