+++ model/sharing/rule_test.go
