+++ model/sharing/setup.go
