+++ model/sharing/shared.go
