+++ web/sharings/replicator_test.go
@@ -9,11 +9,11 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/cozy/cozy-stack/model/sharing"
 	"github.com/cozy/cozy-stack/pkg/consts"
 	"github.com/cozy/cozy-stack/pkg/couchdb"
 	"github.com/cozy/cozy-stack/pkg/instance"
 	"github.com/cozy/cozy-stack/pkg/permissions"
-	"github.com/cozy/cozy-stack/pkg/sharing"
 	"github.com/cozy/cozy-stack/pkg/vfs"
 	"github.com/cozy/echo"
 	uuid "github.com/satori/go.uuid"
